drm/i915: Wrap plane and crtc state to begin separation of uapi and hw state
With some of the newer hardware capabilities coming down the pipeline
(gen11+ NV12, gen11+ big joiner) we wind up with situations where the
state that we use to program a mode object is not based on the state
provided via the uapi (i.e., we secretly steal slave planes or CRTCs
without userspace knowing).  As we move down this path, it's important
to keep the uapi state requested by userspace distinct from the state
that's actually being used to program the hardware.

This is a pretty invasive overhaul, so to minimize the code churn as
much as possible, our intel_{obj}_state structures are now considered
the "hardware state" for an object and most of our driver code will
continue to operate on these as it always has.  Note that
intel_obj_state->base is no longer the true base class that the DRM core
operates on, but rather a driver-internal copy of that structure that
we're free to modify without impacting uapi.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>