drm/i915: Track success of combo phy verification properly
During combo phy programming verification we track the overall
success/failure in a boolean 'ret' variable.  However we accidentally
added a new step at the beginning that performs an &= on the
uninitialized ret, and then later gets its result clobbered by a direct
assignment on what used to be the first step.

Fixes: 239bef676d8e ("drm/i915/display: Implement new combo phy initialization step")
Cc: José Roberto de Souza <jose.souza@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>