drm/i915: Convert INTEL_GEN() to DISPLAY_VER() as appropriate in intel_pm.c
Although most of the code in this file is display-related (watermarks),
there's some functions that are not (e.g., clock gating).  Thus we need
to do the conversions to DISPLAY_VER() manually here rather than using
Coccinelle.

In the near-future we'll probably want to think about moving watermark
logic out of intel_pm.c and into watermark-specific files under the
display/ directory.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>