drm/xe: Move 'vm_max_level' flag back to platform descriptor
The number of page table levels for PPGTT virtual addresses is something
that should be tracked at the platform level rather than the IP level.
Even when mixing and matching various graphics, media, and display IP
blocks, the platform as a whole has to have consistent page table
handling.  This is also a trait that should be tied to the platform even
if the graphics IP itself is not present (e.g., if we disable the
primary GT via configfs).

Since all platforms officially supported by the Xe driver use the same
value (4 levels) and only some old Xe1 platforms have a different value,
let's make this the default if a platform descriptor doesn't specify the
value.  If more platforms show up in the future with different numbers
of page table levels, we can go back to requiring this to be explicitly
specified for every platform at that time.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>