drm/xe: Move 'va_bits' flag back to platform descriptor
The number of virtual address bits is something that should be tracked
at the platform level rather than the IP level.  Even when mixing and
matching various graphics, media, and display IP blocks, the platform as
a whole has to have consistent page table handling.  This is also a
trait that should be tied to the platform even if the graphics IP itself
is not present (e.g., if we disable the primary GT via configfs).

Since all platforms except PVC use the same value (48 bits), let's make
this the default if a platform descriptor doesn't specify the value.  If
more platforms show up in the future with virtual addresses of 57-bits
(or other sizes), we can go back to requiring this to be explicitly
specified for every platform at that time.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>