drm/xe/forcewake: Create dedicated type for forcewake references
xe_force_wake_get() currently returns an integer mask of power domains
that were successfully awoken; both this mask and a pointer to the force
wake collection must be passed to xe_force_wake_put() to release the
wake reference.

Create a dedicated structure type to hold both the mask and the
collection pointer.  While this change does little on its own, it will
make it easier for us to add scope-based cleanup of forcewake in the
future.

FIXME:
   For ease of review, this patch contains only the manual changes to
   add the structure and change the get/put function definitions; it
   does not build on its own since the rest of the driver is still
   trying to call the get/put functions with the old signature.  The
   next patch contains the coccinelle-generated changes necessary
   elsewhere in the driver to adapt to the new interface.  The two
   patches will be squashed together when applied, remain separate for
   now to help reviewers.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>