squash! squash! drm/xe/forcewake: Create dedicated type for forcewake references
Additional xe_force_wake_ref conversions that are unusual and thus
weren't picked up automatically by Coccinelle and needed to be adjusted
by hand.  Most of these are cases where 'fw_ref' was declared on the
same line as other unsigned int variables (I'm sure there's a proper way
to split it out automatically in smpl, but I couldn't figure it out).

This will be squashed into the conversion patch before applying.  It
remains separate now to make the conversion easier to review.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>