drm/xe: Create scoped cleanup class for force_wake_get_any_engine()
force_wake_get_any_engine() is a single-use function to pick any engine
present on the platform and grab its forcewake.  The signature
(returning a boolean success and both the engine pointer and a forcewake
ref by reference) is a bit awkward.  Rewrite it such that the
forcewake ref is the function's return value and the caller can
determine success/failure by checking the engine pointer against NULL.

With this new signature, the function can serve as a scoped cleanup
class constructor, so define the corresponding class.  Note that if we
fail to obtain forcewake (or if the platform somehow has no engines),
the constructor can fail, returning an invalid fw_ref.  In such cases,
fw_ref.fw will be NULL, making it clear that the reference is invalid;
this fact can be used to create a thin wrapper around xe_force_wake_put
that can be used as a destructor for this class.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>